Created Leaderboard and answerResults screen
answerResult receives an correct parameter (if they got the drawing right or not) as well as the description of the specific drawing. Leaderboard contains position (all this stuff will be changed and stored locally). Make sure to change the href on the button at the bottom to redirect to main page once main page is made.